Solution for Day 3 Part 2.
Here we stored the sequences of numbers in a map with keys corresponds to the coords and a unique id. Another loop is to locate the * sign, then get the neighboring numbers. We used set to ensure uniqueness of neighbour. If there are 2 neighbours we can then compute the gear sets and their summations.